Ball, physics, cup added
ball added and code that when you click on the ball it pushes it forward if you hold click it goes further. use keys a and d to rotate the ball and added bounciness and a cup that takes you to the next scene. problem im having is the rotation of the ball on the x axis not freezing causing the ball to jump when hit.